+++ modules/Students/includes/PeopleFields.php
@@ -40,7 +40,7 @@ if ( isset( $_POST['tables'] )
 						$_REQUEST['category_id'] = $columns['CATEGORY_ID'];
 					}
 
-					$sql = 'UPDATE ' . $table . ' SET ';
+					$sql = 'UPDATE ' . DBEscapeIdentifier( $table ) . ' SET ';
 
 					foreach ( (array) $columns as $column => $value )
 					{
@@ -54,7 +54,7 @@ if ( isset( $_POST['tables'] )
 				// New Field / Category.
 				else
 				{
-					$sql = 'INSERT INTO ' . $table . ' ';
+					$sql = 'INSERT INTO ' . DBEscapeIdentifier( $table ) . ' ';
 
 					// New Field.
 					if ( $table === 'PEOPLE_FIELDS' )